example/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-core==2.3.2
google-cloud-secret-manager==2.16.0
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.51.3
idna==3.4
multidict==6.0.4
openai==0.27.2
Pillow==9.4.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.28.2
rsa==4.9
six==1.16.0
tqdm==4.65.0
urllib3==1.26.15
yarl==1.8.2
